싱글톤관리 객체 추가
