+++ Assets/Scripts/DirectionalLight.cs
@@ -1,0 +1,8 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class DirectionalLight : Singleton<DirectionalLight>
+{
+
+}
