+++ Assets/Scripts/Managers/ObtacleManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b3f5879cf10a6424d9caf47a5fb6da69
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
