+++ Assets/Scripts/Utils/CameraController.cs
@@ -1,0 +1,22 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class CameraController : Singleton<CameraController>
+{
+    float rotX;
+    float rotY;
+    float sensitivity = 400f;
+    float clampAngle = 40f;
+
+    void Update()
+    {
+        // 카메라의 회전
+        rotX += -(Input.GetAxis("Mouse Y")) * sensitivity * Time.deltaTime;
+        rotY += Input.GetAxis("Mouse X") * sensitivity * Time.deltaTime;
+
+        rotX = Mathf.Clamp(rotX, -clampAngle, clampAngle);
+        Quaternion rot = Quaternion.Euler(rotX, rotY, 0);
+        transform.rotation = rot;
+    }
+}
