--- Assets/Scripts/Utils/CameraRotation.cs
@@ -1,54 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class CameraRotation : MonoBehaviour
-{
-    float rotX;
-    float rotY;
-    float sensitivity = 400f;
-    float clampAngle = 40f;
-    private static CameraRotation _instance;
-
-    public static CameraRotation Instance
-    {
-        get
-        {
-            // 인스턴스가 없는 경우에 접근하려 하면 인스턴스를 할당.
-            if (!_instance)
-            {
-                _instance = FindObjectOfType(typeof(CameraRotation)) as CameraRotation;
-
-                if (_instance == null)
-                    Debug.Log("no Singleton obj");
-            }
-            return _instance;
-        }
-        set
-        {
-            _instance = value;
-        }
-    }
-
-    void Awake()
-    {
-        if (Instance == null)
-            Instance = this;
-        // 인스턴스가 존재하는 경우 새로생기는 인스턴스를 삭제한다.
-        else if (Instance != this)
-            Destroy(gameObject);
-        // 아래의 함수를 사용하여 씬이 전환되더라도 선언되었던 인스턴스가 파괴되지 않는다.
-    }
-
-
-    void Update()
-    {
-        // 카메라의 회전
-        rotX += -(Input.GetAxis("Mouse Y")) * sensitivity * Time.deltaTime;
-        rotY += Input.GetAxis("Mouse X") * sensitivity * Time.deltaTime;
-
-        rotX = Mathf.Clamp(rotX, -clampAngle, clampAngle);
-        Quaternion rot = Quaternion.Euler(rotX, rotY, 0);
-        transform.rotation = rot;
-    }
-}
